backend/requirements.txt
fastapi>=0.115.0
pydantic[email]>=2.0
pydantic-settings>=2.0
mangum>=0.19.0
boto3>=1.34.0
bcrypt>=4.1.0
PyJWT>=2.8.0
httpx>=0.27.0
pdfplumber>=0.11.0
python-multipart>=0.0.9
uvicorn>=0.30.0
